requirements.txt
numpy
Pillow
PyQt6
matplotlib